Fix memory corruption issues in JsonStructured.
Previously created substrings via (pointer, length) pairs with the
corresponding std::string constructor, this created temporary objects
which we referred into after they had been deallocated. string_view
would probably have been the solution we're looking for, however, we
are running c++11 and can't enjoy the fruit of that labour, therefore
we have a hacky Str class that is private to the JsonStructured class
which solves the mentioned problem.